+++ src/modules/layerInformation/tests/unit/components/LayerInformation.spec.js
@@ -11,11 +11,29 @@ describe("src/modules/layerInformation/components/LayerInformation.vue", () => {
         legendAvailable,
         downloadLinks,
         pointOfContact,
-        publisher;
+        publisher,
+        mainMenu;
 
     beforeEach(() => {
         downloadLinks = null;
         legendAvailable = true;
+        mainMenu = {
+            navigation: {
+                currentComponent: {
+                    type: "layerInformation"
+                }
+            },
+            sections: [
+                [
+                    {
+                        type: "layerInformation"
+                    },
+                    {
+                        type: "contact"
+                    }
+                ]
+            ]
+        };
         store = createStore({
             namespaced: true,
             modules: {
@@ -57,6 +75,26 @@ describe("src/modules/layerInformation/components/LayerInformation.vue", () => {
                             actions: {
                                 createLegendForLayerInfo: sinon.stub()
                             }
+                        }
+                    }
+                },
+                Menu: {
+                    namespaced: true,
+                    getters: {
+                        mainMenu: () => mainMenu,
+                        secondaryMenu: () => {
+                            return {
+                                sections: [
+                                    [
+                                        {
+                                            type: "section1"
+                                        },
+                                        {
+                                            type: "section2"
+                                        }
+                                    ]
+                                ]
+                            };
                         }
                     }
                 }
@@ -189,5 +227,80 @@ describe("src/modules/layerInformation/components/LayerInformation.vue", () => {
         });
 
         expect(wrapper.find("#accordion-container-layer-info-contact").exists()).to.be.true;
+    });
+
+    it("should show contact button", async () => {
+        pointOfContact = {
+            "name": "Behörde ABC",
+            "positionName": ["Metadaten-Verantwortlicher"],
+            "street": "XYZ Straße 99",
+            "housenr": "",
+            "postalCode": "D-12345",
+            "city": "Hamburg",
+            "email": "test@gv.hamburg.de",
+            "country": "DEU"
+        };
+        publisher = "";
+
+        const wrapper = mount(LayerInformationComponent, {
+            global: {
+                plugins: [store]
+            }
+        });
+
+        expect(wrapper.find("#openContactButton").exists()).to.be.true;
+    });
+
+    it("should not show contact button if no contact information is in the metadata", async () => {
+        pointOfContact = "";
+        publisher = "";
+
+        const wrapper = mount(LayerInformationComponent, {
+            global: {
+                plugins: [store]
+            }
+        });
+
+        expect(wrapper.find("#openContactButton").exists()).to.be.false;
+    });
+
+    it("should not show contact button if contact module is not configured", async () => {
+        pointOfContact = {
+            "name": "Behörde ABC",
+            "positionName": ["Metadaten-Verantwortlicher"],
+            "street": "XYZ Straße 99",
+            "housenr": "",
+            "postalCode": "D-12345",
+            "city": "Hamburg",
+            "email": "test@gv.hamburg.de",
+            "country": "DEU"
+        };
+        publisher = "";
+
+        mainMenu = {
+            navigation: {
+                currentComponent: {
+                    type: "layerInformation"
+                }
+            },
+            sections: [
+                [
+                    {
+                        type: "layerInformation"
+                    },
+                    {
+                        type: "about"
+                    }
+                ]
+            ]
+        };
+
+        const wrapper = mount(LayerInformationComponent, {
+            global: {
+                plugins: [store]
+            }
+        });
+
+        expect(wrapper.find("#openContactButton").exists()).to.be.false;
     });
 });